reduce images when downloading them. reduce the size to be at most 500 pixels width or height, grayscale them, and save them as bitmaps (for faster loading).
this reduces the amount of time it takes to search images by an order of magnitude. processing 176 images with a Ryzen 7 2700 (16 cores):
- original: 40 seconds, 90% CPU, 12 GB RAM. 48.5 MB SSD space.
- reduced: 3.9 seconds, 90% CPU, 661 MB RAM. 91.2 MB SSD space.

this is a MASSIVE improvement, but it's still way too inefficient if we're going to service all of ruqqus. next ideas are categorizing images by color (before grayscaling them of course) and matching images using the GPU. There may also be a way to store the features information on disk, so you only have to compute it once for each image.