Storing the entire filestore globally in RAM leads to search times on the order of 70 milliseconds. This is an acceptable speed, it shouldn't take much more than a few seconds to search all of ruqqus like this, unless FileStorage has scaling issues.
I'll still look into GPU FLANN, because that might be fun, and I have the perfect system to test it with.

However, this is essentially a global array and I'd like to replace it with a more maintainable solution. We also need to save it to the disk regularly, and it needs to deal with concurrent access. When this becomes public and if it becomes popular, we will NEED to make sure it can handle many concurrent read and write operations.